python/requirements.txt
# /Users/nickfox137/Documents/llm-creative-studio/python/requirements.txt
fastapi
uvicorn
openai
google-generativeai
python-dotenv
anthropic
pypdf
langchain
langchain-community
langchain-openai
langchain-google-genai
langchain-anthropic
langchain-core
pyautogen>=0.2.0
pytest-asyncio>=0.21.0
